+++ packages/analyzer/src/test/static/class/annotated-classes.js
@@ -14,3 +14,13 @@ const hasConstructorAnnotation = HTMLElement;
  * @appliesMixin someMixin
  */
 const hasConstructorExtendsMixinAnnotations = someMixin(HTMLElement);
+
+/**
+ * @private
+ * @constructor
+ * @extends {HTMLElement}
+ * @appliesMixin someMixin
+ */
+const ephemeralSuperclass = someMixin(HTMLElement);
+
+class hasEphemeralSuperclass extends ephemeralSuperclass {}